requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.1
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.0.1
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
gunicorn
whitenoise
dj-database-url
